pyproject.toml
[project]
name = "temporal-daf-poc"
version = "0.1.0"
description = "Temporal Durable Agent Framework POC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioconsole>=0.8.1",
    "flit>=3.12.0",
    "google-adk>=1.1.1",
    "inflection>=0.5.1",
    "slack-sdk>=3.35.0",
    "temporalio>=1.11.1",
    "vertexai>=1.43.0",
]

test = [
  "pytest-asyncio>=0.25.0",
  "pytest-mock>=3.14.0",
  "pytest-xdist>=3.6.1",
  "pytest>=8.3.4"
]

[tool.flit.sdist]
include = ['src/**/*', 'README.md', 'pyproject.toml', 'LICENSE']
exclude = ['src/**/*.sh']

[tool.flit.module]
name = "temporal.agent"

[build-system]
# Build system specify which backend is used to build/install the project (flit,
# poetry, setuptools,...). All backends are supported by `pip install`
requires = ["flit_core >=3.8,<4"]
build-backend = "flit_core.buildapi"
